+++ pytest_qaseio/api_client.py
@@ -1,8 +1,9 @@
 import logging
 import sys
+from typing import cast
 
 import qaseio
-import qaseio.apis
+from qaseio import configuration as qaseio_config
 from qaseio import models
 
 
@@ -21,7 +22,7 @@ class QaseClient:
             logging.StreamHandler(sys.stderr),
         )
         self._client = qaseio.ApiClient(
-            configuration=qaseio.Configuration(
+            configuration=qaseio_config.Configuration(
                 api_key={
                     "TokenAuth": token,
                 },
@@ -33,20 +34,24 @@ class QaseClient:
         self,
         run_id: int,
     ) -> models.Run:
-        return qaseio.apis.RunsApi(self._client).get_run(
-            code=self._project_code, id=run_id,
-        ).result
+        run_response = qaseio.RunsApi(self._client).get_run(
+            code=self._project_code,
+            id=run_id,
+        )
+        return cast(models.Run, run_response.result)
 
     def create_run(
         self,
         run_data: models.RunCreate,
     ) -> models.Run:
         """Create test run in Qase."""
-        response = qaseio.apis.RunsApi(self._client).create_run(
+        response = qaseio.RunsApi(self._client).create_run(
             code=self._project_code,
             run_create=run_data,
         )
-        return self.get_run(response.result.id)
+        created_run = cast(models.IdResponseAllOfResult, response.result)
+
+        return self.get_run(cast(int, created_run.id))
 
     def load_cases_ids(
         self,
@@ -54,18 +59,19 @@ class QaseClient:
         """Load all cases ids of project."""
         limit = 100
         cases: list[int] = []
-        response: models.TestCaseListResponse = qaseio.apis.CasesApi(self._client).get_cases(
+        response = qaseio.CasesApi(self._client).get_cases(
             code=self._project_code,
             limit=limit,
             offset=len(cases),
         )
         while True:
-            response = qaseio.apis.CasesApi(self._client).get_cases(
+            response = qaseio.CasesApi(self._client).get_cases(
                 code=self._project_code,
                 limit=limit,
                 offset=len(cases),
             )
-            new_cases = [case.id for case in response.result.entities]
+            response_cases = getattr(response.result, "entities", [])
+            new_cases = [case.id for case in response_cases]
             cases += new_cases
             if not new_cases:
                 break
@@ -73,13 +79,18 @@ class QaseClient:
 
     def report_test_results(
         self,
-        run: models.RunCreate,
+        run: models.Run,
         report_data: models.ResultCreate,
     ) -> tuple[str, models.ResultCreate]:
         """Report test results back to Qase."""
-        result = qaseio.apis.ResultsApi(self._client).create_result(
-            code=self._project_code,
-            id=run.id,
-            result_create=report_data,
-        ).result
-        return result.hash, report_data.status
+        result = (
+            qaseio.ResultsApi(self._client)
+            .create_result(
+                code=self._project_code,
+                id=cast(int, run.id),
+                result_create=report_data,
+            )
+            .result
+        )
+        assert result
+        return cast(str, result.hash), cast(models.ResultCreate, report_data.status)