+++ pytest_qaseio/storage.py
@@ -1,8 +1,8 @@
 import io
-from typing import Protocol
+from typing import Protocol, cast
 
 import qaseio
-import qaseio.apis
+from qaseio import configuration as qaseio_config
 
 
 class FileStorage(Protocol):
@@ -10,6 +10,22 @@ class FileStorage(Protocol):
 
     def save_file_obj(self, content: bytes, filename: str) -> str:
         """Upload file to storage and return URL."""
+        ...
+
+
+class FileIO(io.BytesIO):
+    """Represent file object to pass in Qase API methods.
+
+    Set `name` and `mime` attributes to objects since it's required in
+    https://github.com/qase-tms/qase-python/blob/44d19a500246017a30e0fa06b35cace065135d96/qaseio/src/qaseio/api_client.py#L518
+
+    """
+
+    mime = "image/png"
+
+    def __init__(self, *args, filename: str, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+        self.name = filename
 
 
 class QaseFileStorage:
@@ -22,7 +38,7 @@ class QaseFileStorage:
     ):
         """Prepare ApiClient for qase io using credentials."""
         self._client = qaseio.ApiClient(
-            configuration=qaseio.Configuration(
+            configuration=qaseio_config.Configuration(
                 api_key={
                     "TokenAuth": qase_token,
                 },
@@ -32,11 +48,18 @@ class QaseFileStorage:
 
     def save_file_obj(self, content: bytes, filename: str) -> str:
         """Upload file to Qase.io S3 bucket via attachment API."""
-        file_obj = io.BytesIO(content)
-        file_obj.name = filename
+        file_obj = FileIO(content, filename=filename)
 
-        result = qaseio.apis.AttachmentsApi(self._client).upload_attachment(
-            code=self._project_code,
-            file=[file_obj],
-        ).result[0]
-        return result["url"]
+        attachment_response_result = (
+            qaseio.AttachmentsApi(
+                self._client,
+            )
+            .upload_attachment(
+                code=self._project_code,
+                file=[file_obj],
+            )
+            .result
+        )
+        assert attachment_response_result is not None
+
+        return cast(str, attachment_response_result[0].url)