+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 We follow [Semantic Versions](https://semver.org/).
 
+# Version 2.0.0
+
+- Use updated `qaseio` sdk (^4.0.0)
+- Rework getting browser name (see `README.md`)
+- Drop `Python 3.10` support
+- Drop `pytest-variables` support (now use pytest options to provide browser name)
+- Add `ruff` and `cspell` linters
+- Small code and structure improvements
+
 # Version 1.2.0
 
 - Fix `filelock` blocking for parallel tests run